// src/types/Servicios.ts
export default interface Servicios {
  idServicio: string;
  nombre: string;
}

// agua?: boolean;
// luz?: boolean;
// gas?: boolean;
// telefono?: boolean;
// internet?: boolean;
// cable?: boolean;
// cisterna?: boolean;
// calentador?: boolean;
// aireAcondicionado?: boolean;
// calefaccion?: boolean;
// elevador?: boolean;
// seguridad?: boolean;
// alarma?: boolean;
// plantaElectrica?: boolean;
// otros?: string;