// src/types/Amenidades.ts
export default interface Amenidades {
  idAmenidad?: string;
  nombre?: string;
}

// alberca?: boolean;
// jardin?: boolean;
// terraza?: boolean;
// balcon?: boolean;
// gimnasio?: boolean;
// asador?: boolean;
// areaJuegosInfantiles?: boolean;
// salonDeFiestas?: boolean;
// campoGolf?: boolean;
// otros?: string;